backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Pydantic v2
pydantic==2.5.3
pydantic-settings==2.1.0

# Async HTTP client
aiohttp==3.9.1

# Configuration
pyyaml==6.0.1

# Database (Async SQLAlchemy + SQLite)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dateutil==2.8.2

# Background task scheduler
apscheduler==3.10.4

# Media name parsing
parse-torrent-title==2.8.1
